hay un error con la grafica

+++ py5_matplotlib/plot_list.py
@@ -1,6 +1,8 @@
+import matplotlib
 import matplotlib.pyplot as plt
 
 ndelista = int(input('Escriba el numero para la lista:  '))
+
 num = ndelista + 10;
 print('El número escogido más 10 es: ',num)
 dos = 2;
@@ -8,7 +10,7 @@ resu = [];
 while num > 1 :
 
  if (num % dos) == 0 :
-  num = num / 2;
+  num = int(num / 2);
  elif (num % dos) == 1 :
   num = (num*3) + 1;
